Extract workspace_ops, input, notifications_ui, startup, and fonts modules (#68)
Phase 2-3 of main.rs decomposition. Moves ~2400 lines from main.rs into
5 focused modules while keeping the AmuxApp struct and core update loop
in main.rs. main.rs drops from 4114 to 1758 lines.

New modules:
- workspace_ops.rs (376 lines): pane/workspace lifecycle, split, close, navigate, scroll
- input.rs (920 lines): keyboard shortcuts, copy-mode, selection mouse, clipboard
- notifications_ui.rs (364 lines): notification delivery, bubbling, panel rendering
- startup.rs (527 lines): app bootstrap, session restore, PTY spawning
- fonts.rs (197 lines): system font fallback loading, bold font helper
