+++ crates/amux-app/src/main.rs
@@ -1,12 +1,17 @@
+mod fonts;
+mod input;
 mod ipc_dispatch;
 mod key_encode;
 mod managed_pane;
 mod menu_bar;
+mod notifications_ui;
 mod render;
 mod selection;
 mod sidebar;
+mod startup;
 mod system_notify;
 mod theme;
+mod workspace_ops;
 
 use std::collections::HashMap;
 use std::io::Read;
@@ -82,726 +87,8 @@ const TERMINAL_TOP_PAD: f32 = 28.0;
 /// Painted with terminal background color so it blends with the terminal.
 const TERMINAL_BOTTOM_PAD: f32 = 4.0;
 
-// ---------------------------------------------------------------------------
-// App config (loaded from ~/.config/amux/config.toml)
-// ---------------------------------------------------------------------------
-
-/// Cached font family for semibold/bold UI text (sidebar titles, active tabs).
-/// Uses a static Arc<str> to avoid allocating on every call.
-pub(crate) fn bold_font(size: f32) -> egui::FontId {
-    use std::sync::LazyLock;
-    static BOLD_FAMILY: LazyLock<egui::FontFamily> =
-        LazyLock::new(|| egui::FontFamily::Name("Bold".into()));
-    egui::FontId::new(size, BOLD_FAMILY.clone())
-}
-
-/// Load system fonts as fallbacks to egui's font families.
-/// This provides coverage for braille patterns, geometric shapes, and other symbols
-/// that egui's bundled Hack font doesn't include.
-///
-/// Also registers bundled IBM Plex Sans (Regular + SemiBold) for consistent
-/// cross-platform UI text, and a custom "Bold" font family for titles.
-fn install_system_font_fallback(ctx: &egui::Context) {
-    let mut fonts = egui::FontDefinitions::default();
-    let mut loaded = Vec::new();
-
-    // --- Bundled IBM Plex Sans for UI chrome ---
-    fonts.font_data.insert(
-        "plex_sans".to_owned(),
-        egui::FontData::from_static(font::SANS_REGULAR).into(),
-    );
-    fonts.font_data.insert(
-        "plex_sans_semibold".to_owned(),
-        egui::FontData::from_static(font::SANS_SEMIBOLD).into(),
-    );
-
-    // Platform-specific font candidates: (path, name, is_symbol_font, is_proportional)
-    // We try to load a monospace font + a symbols font for maximum coverage.
-    let candidates: &[(&str, &str, bool, bool)] = if cfg!(target_os = "macos") {
-        &[
-            // SF Pro (system font) for UI chrome text
-            ("/System/Library/Fonts/SFNS.ttf", "sf_pro", false, true),
-            // SF Mono: single .ttf with good Unicode coverage
-            (
-                "/System/Library/Fonts/SFNSMono.ttf",
-                "sf_mono",
-                false,
-                false,
-            ),
-            // Apple Symbols: broad Unicode symbol coverage
-            (
-                "/System/Library/Fonts/Apple Symbols.ttf",
-                "apple_symbols",
-                true,
-                false,
-            ),
-            // Supplemental Andale Mono as another option
-            (
-                "/System/Library/Fonts/Supplemental/Andale Mono.ttf",
-                "andale_mono",
-                false,
-                false,
-            ),
-        ]
-    } else if cfg!(target_os = "windows") {
-        &[
-            ("C:\\Windows\\Fonts\\segoeui.ttf", "segoe_ui", false, true),
-            ("C:\\Windows\\Fonts\\consola.ttf", "consolas", false, false),
-            (
-                "C:\\Windows\\Fonts\\segmdl2.ttf",
-                "segoe_symbols",
-                true,
-                false,
-            ),
-        ]
-    } else {
-        &[
-            (
-                "/usr/share/fonts/truetype/dejavu/DejaVuSansMono.ttf",
-                "dejavu_mono",
-                false,
-                false,
-            ),
-            (
-                "/usr/share/fonts/TTF/DejaVuSansMono.ttf",
-                "dejavu_mono",
-                false,
-                false,
-            ),
-        ]
-    };
-
-    // Bold system font candidates: (path, name)
-    // Loaded separately so they can be added to the "Bold" family.
-    let bold_candidates: &[(&str, &str)] = if cfg!(target_os = "macos") {
-        // macOS provides bold weights / variable font axes for SF Pro, but we
-        // intentionally avoid hard-coding those paths or relying on axis selection.
-        // Bundled Plex Sans SemiBold is used instead for bold UI text.
-        &[]
-    } else if cfg!(target_os = "windows") {
-        &[("C:\\Windows\\Fonts\\segoeuib.ttf", "segoe_ui_bold")]
-    } else {
-        &[(
-            "/usr/share/fonts/truetype/dejavu/DejaVuSans-Bold.ttf",
-            "dejavu_bold",
-        )]
-    };
-
-    let mut proportional_fonts = Vec::new();
-    let mut mono_fonts = Vec::new();
-    let mut bold_fonts = Vec::new();
-
-    for &(path, name, _is_symbol, is_proportional) in candidates {
-        if fonts.font_data.contains_key(name) {
-            continue;
-        }
-        match std::fs::read(path) {
-            Ok(data) => {
-                fonts
-                    .font_data
-                    .insert(name.to_owned(), egui::FontData::from_owned(data).into());
-                loaded.push(name);
-                if is_proportional {
-                    proportional_fonts.push(name);
-                } else {
-                    mono_fonts.push(name);
-                }
-            }
-            Err(e) => {
-                tracing::debug!("Font fallback not found: {} ({})", path, e);
-            }
-        }
-    }
-
-    for &(path, name) in bold_candidates {
-        if fonts.font_data.contains_key(name) {
-            continue;
-        }
-        match std::fs::read(path) {
-            Ok(data) => {
-                fonts
-                    .font_data
-                    .insert(name.to_owned(), egui::FontData::from_owned(data).into());
-                bold_fonts.push(name);
-                loaded.push(name);
-            }
-            Err(e) => {
-                tracing::debug!("Bold font not found: {} ({})", path, e);
-            }
-        }
-    }
-
-    // Proportional fonts: bundled Plex Sans first, then system UI font, then mono
-    if let Some(family) = fonts.families.get_mut(&egui::FontFamily::Proportional) {
-        family.insert(0, "plex_sans".to_owned());
-        for name in &proportional_fonts {
-            family.insert(1, (*name).to_owned());
-        }
-        for name in &mono_fonts {
-            family.push((*name).to_owned());
-        }
-    }
-    // Monospace fonts: add all as fallbacks
-    if let Some(family) = fonts.families.get_mut(&egui::FontFamily::Monospace) {
-        for name in &loaded {
-            family.push((*name).to_owned());
-        }
-    }
-
-    // "Bold" family: bundled Plex Sans SemiBold first, then system bold fonts,
-    // then fall back to all proportional + mono fonts for symbol coverage.
-    let mut bold_family = vec!["plex_sans_semibold".to_owned()];
-    for name in &bold_fonts {
-        bold_family.push((*name).to_owned());
-    }
-    for name in &proportional_fonts {
-        bold_family.push((*name).to_owned());
-    }
-    for name in &mono_fonts {
-        bold_family.push((*name).to_owned());
-    }
-    // Include egui's bundled fonts as final fallback
-    if let Some(prop) = fonts.families.get(&egui::FontFamily::Proportional) {
-        for name in prop {
-            if !bold_family.contains(name) {
-                bold_family.push(name.clone());
-            }
-        }
-    }
-    fonts
-        .families
-        .insert(egui::FontFamily::Name("Bold".into()), bold_family);
-
-    if loaded.is_empty() {
-        tracing::warn!("No system font fallbacks found; egui may miss symbol/emoji coverage");
-    }
-
-    tracing::info!("Loaded font fallbacks: {:?}", loaded);
-    ctx.set_fonts(fonts);
-}
-
 fn main() -> anyhow::Result<()> {
-    tracing_subscriber::fmt::init();
-
-    let app_config = config::load_app_config();
-    let font_size = app_config.font_size;
-    // FontConfig is only consumed by the GPU renderer; gate to avoid unused
-    // warnings in non-GPU builds. font_family is GPU-only — the egui fallback
-    // renderer uses its built-in monospace font.
-    #[cfg(feature = "gpu-renderer")]
-    let font_config = font::FontConfig {
-        family: app_config.font_family.clone(),
-        size: app_config.font_size,
-    };
-
-    // Initialize sound player with configured sound setting
-    let mut sound_player = system_notify::SoundPlayer::new();
-    if let Some(player) = &mut sound_player {
-        player.configure(&app_config.notifications.sound.sound);
-    }
-
-    let socket_token = uuid::Uuid::new_v4().to_string();
-    let (ipc_rx, ipc_addr, event_broadcaster) = amux_ipc::start_server(socket_token.clone())?;
-    tracing::info!("IPC server: {}", ipc_addr);
-
-    let theme = theme::Theme::default();
-    let mut term_config = AmuxTermConfig::default();
-    theme.apply_to_palette(&mut term_config.color_palette);
-    let config = Arc::new(term_config);
-
-    // Try to restore a previous session
-    let restored = match amux_session::load() {
-        Ok(Some(session)) => {
-            tracing::info!("Restoring session from {}", session.saved_at);
-            Some(session)
-        }
-        Ok(None) => None,
-        Err(amux_session::SessionError::VersionMismatch { version, expected }) => {
-            tracing::warn!(
-                "Session version {} not supported (expected {}), starting fresh",
-                version,
-                expected
-            );
-            None
-        }
-        Err(amux_session::SessionError::Corrupted(e)) => {
-            tracing::error!("Session file corrupted: {}, starting fresh", e);
-            None
-        }
-        Err(e) => {
-            tracing::warn!("Failed to load session, starting fresh: {}", e);
-            None
-        }
-    };
-
-    let state = if let Some(session) = restored {
-        restore_session(&session, &ipc_addr, &socket_token, &config)
-    } else {
-        fresh_startup(&ipc_addr, &socket_token, &config)?
-    };
-
-    // Force dark appearance on macOS so the title bar matches the app's dark chrome.
-    #[cfg(target_os = "macos")]
-    {
-        use objc2_app_kit::{NSAppearance, NSApplication};
-        use objc2_foundation::{MainThreadMarker, NSString};
-
-        if let Some(mtm) = MainThreadMarker::new() {
-            let app = NSApplication::sharedApplication(mtm);
-            let dark =
-                NSAppearance::appearanceNamed(&NSString::from_str("NSAppearanceNameDarkAqua"));
-            app.setAppearance(dark.as_deref());
-        }
-    }
-
-    // Build the native menu bar (cross-platform via muda).
-    let menu = menu_bar::build();
-
-    let viewport = egui::ViewportBuilder::default()
-        .with_inner_size([1000.0, 600.0])
-        .with_title("amux");
-    #[cfg(target_os = "macos")]
-    let viewport = viewport
-        .with_fullsize_content_view(true)
-        .with_titlebar_shown(false)
-        .with_title_shown(false);
-
-    let options = eframe::NativeOptions {
-        viewport,
-        // Suppress winit's default macOS menu (we provide our own via muda).
-        #[cfg(target_os = "macos")]
-        event_loop_builder: Some(Box::new(|builder| {
-            use winit::platform::macos::EventLoopBuilderExtMacOS;
-            builder.with_default_menu(false);
-        })),
-        ..Default::default()
-    };
-
-    let ipc_addr_cleanup = ipc_addr.clone();
-    let result = eframe::run_native(
-        "amux",
-        options,
-        Box::new(move |_cc| {
-            // Add system monospace font as fallback for braille/symbol coverage
-            install_system_font_fallback(&_cc.egui_ctx);
-
-            // Hide the panel resize handle entirely (cursor still changes on hover).
-            _cc.egui_ctx.style_mut(|style| {
-                style.visuals.widgets.noninteractive.fg_stroke.color = egui::Color32::TRANSPARENT;
-                style.visuals.widgets.inactive.fg_stroke.color = egui::Color32::TRANSPARENT;
-                style.visuals.widgets.hovered.fg_stroke.color = egui::Color32::TRANSPARENT;
-                style.visuals.widgets.active.fg_stroke.color = egui::Color32::TRANSPARENT;
-            });
-
-            #[cfg(feature = "gpu-renderer")]
-            let gpu_renderer = _cc.wgpu_render_state.as_ref().map(|rs| {
-                tracing::info!("GPU renderer initialized (wgpu backend)");
-                GpuRenderer::new(rs.clone(), &font_config)
-            });
-
-            Ok(Box::new(AmuxApp {
-                workspaces: state.workspaces,
-                active_workspace_idx: state.active_workspace_idx,
-                panes: state.panes,
-                next_pane_id: state.next_pane_id,
-                next_workspace_id: state.next_workspace_id,
-                next_surface_id: state.next_surface_id,
-                sidebar: state.sidebar,
-                ipc_rx,
-                event_broadcaster,
-                socket_addr: ipc_addr,
-                socket_token,
-                config,
-                theme,
-                last_panel_rect: None,
-                notifications: state.notifications,
-                show_notification_panel: false,
-                last_click_time: Instant::now(),
-                last_click_pos: egui::Pos2::ZERO,
-                click_count: 0,
-                wants_exit: false,
-                font_size,
-                find_state: None,
-                copy_mode: None,
-                hovered_hyperlink: None,
-                ime_preedit: None,
-                selection_changed: false,
-                tab_drag: None,
-                rename_modal: None,
-                app_focused: true,
-                app_config,
-                system_notifier: system_notify::SystemNotifier::new(),
-                last_badge_count: 0,
-                sound_player,
-                menu,
-                #[cfg(target_os = "windows")]
-                menu_attached: false,
-                #[cfg(feature = "gpu-renderer")]
-                gpu_renderer,
-            }))
-        }),
-    )
-    .map_err(|e| anyhow::anyhow!("{}", e));
-
-    cleanup_addr(&ipc_addr_cleanup);
-    result
-}
-
-/// Bundled startup state to avoid complex return tuples.
-struct StartupState {
-    workspaces: Vec<Workspace>,
-    active_workspace_idx: usize,
-    panes: HashMap<PaneId, ManagedPane>,
-    next_pane_id: PaneId,
-    next_workspace_id: u64,
-    next_surface_id: u64,
-    sidebar: SidebarState,
-    notifications: NotificationStore,
-}
-
-/// Create a fresh default startup (one workspace, one pane).
-fn fresh_startup(
-    ipc_addr: &amux_ipc::IpcAddr,
-    socket_token: &str,
-    config: &Arc<AmuxTermConfig>,
-) -> anyhow::Result<StartupState> {
-    let initial_pane_id: PaneId = 0;
-    let surface = spawn_surface(80, 24, ipc_addr, socket_token, config, 0, 0, None, None)?;
-
-    let managed = ManagedPane {
-        surfaces: vec![surface],
-        active_surface_idx: 0,
-        selection: None,
-    };
-
-    let mut panes = HashMap::new();
-    panes.insert(initial_pane_id, managed);
-
-    let workspace = Workspace {
-        id: 0,
-        title: "Terminal 1".to_string(),
-        tree: PaneTree::new(initial_pane_id),
-        focused_pane: initial_pane_id,
-        zoomed: None,
-        dragging_divider: None,
-        last_pane_sizes: HashMap::new(),
-        color: None,
-    };
-
-    Ok(StartupState {
-        workspaces: vec![workspace],
-        active_workspace_idx: 0,
-        panes,
-        next_pane_id: 1,
-        next_workspace_id: 1,
-        next_surface_id: 1,
-        sidebar: SidebarState {
-            visible: true,
-            width: DEFAULT_SIDEBAR_WIDTH,
-            drag: None,
-        },
-        notifications: NotificationStore::new(),
-    })
-}
-
-/// Restore app state from a saved session. Falls back to fresh startup on any failure.
-fn restore_session(
-    session: &SessionData,
-    ipc_addr: &amux_ipc::IpcAddr,
-    socket_token: &str,
-    config: &Arc<AmuxTermConfig>,
-) -> StartupState {
-    let mut workspaces = Vec::new();
-    let mut panes: HashMap<PaneId, ManagedPane> = HashMap::new();
-
-    for saved_ws in &session.workspaces {
-        for (&pane_id, saved_pane) in &saved_ws.panes {
-            if saved_pane.panel_type != amux_session::PANEL_TYPE_TERMINAL {
-                tracing::warn!(
-                    "Skipping pane {} with unsupported panel type {:?}",
-                    pane_id,
-                    saved_pane.panel_type,
-                );
-                continue;
-            }
-            let mut surfaces = Vec::new();
-            for saved_sf in &saved_pane.surfaces {
-                let cwd = saved_sf.working_dir.as_deref();
-                let scrollback = if saved_sf.scrollback.is_empty() {
-                    None
-                } else {
-                    Some(saved_sf.scrollback.as_str())
-                };
-
-                match spawn_surface(
-                    saved_sf.cols,
-                    saved_sf.rows,
-                    ipc_addr,
-                    socket_token,
-                    config,
-                    saved_ws.id,
-                    saved_sf.id,
-                    cwd,
-                    scrollback,
-                ) {
-                    Ok(mut surface) => {
-                        // Restore git/PR metadata from session
-                        surface.metadata.git_branch = saved_sf.git_branch.clone();
-                        surface.metadata.git_dirty = saved_sf.git_dirty;
-                        surface.metadata.pr_number = saved_sf.pr_number;
-                        surface.metadata.pr_title = saved_sf.pr_title.clone();
-                        surface.metadata.pr_state = saved_sf.pr_state.clone();
-                        surface.user_title = saved_sf.user_title.clone();
-                        surfaces.push(surface);
-                    }
-                    Err(e) => {
-                        tracing::warn!(
-                            "Failed to restore surface {} in pane {}: {}",
-                            saved_sf.id,
-                            pane_id,
-                            e
-                        );
-                    }
-                }
-            }
-
-            if surfaces.is_empty() {
-                tracing::warn!("All surfaces failed for pane {}, skipping", pane_id);
-                continue;
-            }
-
-            let active_idx = saved_pane
-                .active_surface_idx
-                .min(surfaces.len().saturating_sub(1));
-            panes.insert(
-                pane_id,
-                ManagedPane {
-                    surfaces,
-                    active_surface_idx: active_idx,
-                    selection: None,
-                },
-            );
-        }
-
-        // Verify all pane IDs in the tree were actually restored
-        let tree_pane_ids = saved_ws.tree.iter_panes();
-        let all_panes_restored = tree_pane_ids.iter().all(|id| panes.contains_key(id));
-        if !all_panes_restored {
-            tracing::warn!(
-                "Skipping workspace {} (tree references missing panes)",
-                saved_ws.title
-            );
-            // Clean up any panes we did restore for this workspace
-            for id in &tree_pane_ids {
-                panes.remove(id);
-            }
-            continue;
-        }
-
-        let focused = if panes.contains_key(&saved_ws.focused_pane) {
-            saved_ws.focused_pane
-        } else {
-            *tree_pane_ids.first().unwrap_or(&0)
-        };
-
-        workspaces.push(Workspace {
-            id: saved_ws.id,
-            title: saved_ws.title.clone(),
-            tree: saved_ws.tree.clone(),
-            focused_pane: focused,
-            zoomed: saved_ws.zoomed.filter(|z| panes.contains_key(z)),
-            dragging_divider: None,
-            last_pane_sizes: HashMap::new(),
-            color: saved_ws.color,
-        });
-    }
-
-    // If nothing restored, fall back to fresh
-    if workspaces.is_empty() {
-        tracing::warn!("Session restore produced no workspaces, starting fresh");
-        return match fresh_startup(ipc_addr, socket_token, config) {
-            Ok(result) => result,
-            Err(e) => {
-                tracing::error!("Fresh startup also failed: {}", e);
-                panic!("Cannot start amux: {}", e);
-            }
-        };
-    }
-
-    let sidebar = SidebarState {
-        visible: session.sidebar.visible,
-        width: session.sidebar.width,
-        drag: None,
-    };
-
-    // Restore only already-read notifications (unread ones are stale — the
-    // agent that created them is gone after restart).
-    let mut store = NotificationStore::new();
-    for saved_n in &session.notifications {
-        if !saved_n.read {
-            continue;
-        }
-        let source = match saved_n.source.as_str() {
-            "toast" => NotificationSource::Toast,
-            "bell" => NotificationSource::Bell,
-            _ => NotificationSource::Cli,
-        };
-        store.push_read(
-            saved_n.workspace_id,
-            saved_n.pane_id,
-            saved_n.surface_id,
-            saved_n.title.clone(),
-            saved_n.subtitle.clone(),
-            saved_n.body.clone(),
-            source,
-        );
-    }
-
-    // Don't restore workspace statuses — agent processes don't survive restart,
-    // so any Active/Waiting state would be stale. They start as Idle implicitly.
-
-    let active_idx = session
-        .active_workspace_idx
-        .min(workspaces.len().saturating_sub(1));
-
-    StartupState {
-        workspaces,
-        active_workspace_idx: active_idx,
-        panes,
-        next_pane_id: session.next_pane_id,
-        next_workspace_id: session.next_workspace_id,
-        next_surface_id: session.next_surface_id,
-        sidebar,
-        notifications: store,
-    }
-}
-
-fn cleanup_addr(addr: &amux_ipc::IpcAddr) {
-    match addr {
-        #[cfg(unix)]
-        amux_ipc::IpcAddr::Unix(path) => {
-            let _ = std::fs::remove_file(path);
-        }
-        #[cfg(windows)]
-        amux_ipc::IpcAddr::NamedPipe(_) => {}
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn spawn_surface(
-    cols: u16,
-    rows: u16,
-    ipc_addr: &amux_ipc::IpcAddr,
-    socket_token: &str,
-    config: &Arc<AmuxTermConfig>,
-    workspace_id: u64,
-    surface_id: u64,
-    cwd: Option<&str>,
-    scrollback: Option<&str>,
-) -> anyhow::Result<PaneSurface> {
-    let shell = shell::default_shell();
-    let mut cmd = CommandBuilder::new(&shell);
-    cmd.env("AMUX_SOCKET_PATH", ipc_addr.to_string());
-    cmd.env("AMUX_SOCKET_TOKEN", socket_token);
-    cmd.env("AMUX_WORKSPACE_ID", workspace_id.to_string());
-    cmd.env("AMUX_SURFACE_ID", surface_id.to_string());
-    cmd.env("TERM", "xterm-256color");
-    cmd.env("TERM_PROGRAM", "amux");
-    cmd.env("TERM_PROGRAM_VERSION", env!("CARGO_PKG_VERSION"));
-
-    // Point AMUX_BIN to the CLI binary so shell integration scripts can invoke it
-    // without relying on PATH (macOS path_helper in /etc/zprofile rebuilds PATH).
-    if let Ok(exe) = std::env::current_exe() {
-        if let Some(exe_dir) = exe.parent() {
-            let cli_bin = exe_dir.join("amux");
-            if cli_bin.exists() {
-                cmd.env("AMUX_BIN", cli_bin.to_string_lossy().as_ref());
-            }
-        }
-    }
-
-    // Prepend amux bin dir (containing claude wrapper) to PATH so hooks are
-    // injected at runtime via --settings, scoped to amux sessions only.
-    if let Some(bin_dir) = shell::ensure_claude_wrapper_dir() {
-        let current_path = std::env::var("PATH").unwrap_or_default();
-        let bin_str = bin_dir.to_string_lossy();
-        if !current_path.split(':').any(|d| d == bin_str.as_ref()) {
-            let sep = if current_path.is_empty() { "" } else { ":" };
-            cmd.env("PATH", format!("{bin_str}{sep}{current_path}"));
-        }
-    }
-
-    // Auto-inject shell integration (matching cmux's ZDOTDIR/PROMPT_COMMAND approach)
-    shell::inject_shell_integration(&shell, &mut cmd);
-
-    let actual_cwd = if let Some(dir) = cwd {
-        let path = std::path::Path::new(dir);
-        if path.is_dir() {
-            cmd.cwd(path);
-            Some(dir.to_string())
-        } else {
-            tracing::warn!("Saved working dir no longer exists: {}", dir);
-            if let Some(home) = dirs::home_dir() {
-                cmd.cwd(&home);
-                Some(home.to_string_lossy().to_string())
-            } else {
-                None
-            }
-        }
-    } else {
-        None
-    };
-
-    let mut pane = TerminalPane::spawn(cols, rows, cmd, config.clone())?;
-
-    // Inject scrollback text before starting the reader thread.
-    // feed_bytes writes directly to the terminal state machine, not through the PTY.
-    if let Some(text) = scrollback {
-        if !text.is_empty() {
-            // Convert \n to \r\n for terminal processing, avoiding extra trailing blank line.
-            let normalized = text.replace("\r\n", "\n").replace('\r', "\n");
-            let trimmed = normalized.trim_end_matches('\n');
-            let buffer = trimmed.replace('\n', "\r\n");
-            if !buffer.is_empty() {
-                pane.feed_bytes(buffer.as_bytes());
-                pane.feed_bytes(b"\r\n");
-            }
-        }
-    }
-
-    let mut reader = pane.take_reader().expect("reader already taken");
-    let (byte_tx, byte_rx) = mpsc::sync_channel::<Vec<u8>>(64);
-
-    thread::spawn(move || {
-        let mut buf = [0u8; 8192];
-        loop {
-            match reader.read(&mut buf) {
-                Ok(0) | Err(_) => break,
-                Ok(n) => {
-                    if byte_tx.send(buf[..n].to_vec()).is_err() {
-                        break;
-                    }
-                }
-            }
-        }
-    });
-
-    // Initialize metadata with the actual CWD used (may differ from saved if dir was removed)
-    let metadata = SurfaceMetadata {
-        cwd: actual_cwd,
-        ..Default::default()
-    };
-
-    Ok(PaneSurface {
-        id: surface_id,
-        pane,
-        byte_rx,
-        scroll_offset: 0,
-        scroll_accum: 0.0,
-        metadata,
-        user_title: None,
-        exited: None,
-    })
+    startup::run()
 }
 
 // --- App ---
@@ -1532,7 +819,7 @@ impl AmuxApp {
 
             let active_idx = managed.active_surface_idx;
             let tab_font = egui::FontId::proportional(11.0);
-            let tab_font_bold = bold_font(11.0);
+            let tab_font_bold = fonts::bold_font(11.0);
             let mut x = tab_rect.min.x + 2.0;
 
             // Get pointer state for hover detection and drag
@@ -1755,7 +1042,7 @@ impl AmuxApp {
                         let ws_id = self.active_workspace().id;
                         let sf_id = self.next_surface_id;
                         self.next_surface_id += 1;
-                        if let Ok(surface) = spawn_surface(
+                        if let Ok(surface) = startup::spawn_surface(
                             80,
                             24,
                             &self.socket_addr,
@@ -2002,943 +1289,6 @@ impl AmuxApp {
         }
     }
 
-    // --- Pane/Workspace management ---
-
-    fn spawn_pane_with_surface(&mut self) -> Option<PaneId> {
-        let ws_id = self.active_workspace().id;
-        let sf_id = self.next_surface_id;
-
-        match spawn_surface(
-            80,
-            24,
-            &self.socket_addr,
-            &self.socket_token,
-            &self.config,
-            ws_id,
-            sf_id,
-            None,
-            None,
-        ) {
-            Ok(surface) => {
-                let pane_id = self.next_pane_id;
-                self.next_pane_id += 1;
-                self.next_surface_id += 1;
-                self.panes.insert(
-                    pane_id,
-                    ManagedPane {
-                        surfaces: vec![surface],
-                        active_surface_idx: 0,
-                        selection: None,
-                    },
-                );
-                Some(pane_id)
-            }
-            Err(e) => {
-                tracing::error!("Failed to spawn pane: {}", e);
-                None
-            }
-        }
-    }
-
-    fn create_workspace(&mut self, title: Option<String>) -> Option<u64> {
-        let ws_id = self.next_workspace_id;
-        let title = title.unwrap_or_else(|| format!("Terminal {}", self.workspaces.len() + 1));
-
-        let sf_id = self.next_surface_id;
-
-        match spawn_surface(
-            80,
-            24,
-            &self.socket_addr,
-            &self.socket_token,
-            &self.config,
-            ws_id,
-            sf_id,
-            None,
-            None,
-        ) {
-            Ok(surface) => {
-                self.next_workspace_id += 1;
-                let pane_id = self.next_pane_id;
-                self.next_pane_id += 1;
-                self.next_surface_id += 1;
-                self.panes.insert(
-                    pane_id,
-                    ManagedPane {
-                        surfaces: vec![surface],
-                        active_surface_idx: 0,
-                        selection: None,
-                    },
-                );
-
-                let workspace = Workspace {
-                    id: ws_id,
-                    title,
-                    tree: PaneTree::new(pane_id),
-                    focused_pane: pane_id,
-                    zoomed: None,
-                    dragging_divider: None,
-                    last_pane_sizes: HashMap::new(),
-                    color: None,
-                };
-
-                self.workspaces.push(workspace);
-                self.active_workspace_idx = self.workspaces.len() - 1;
-                Some(ws_id)
-            }
-            Err(e) => {
-                tracing::error!("Failed to spawn pane for workspace: {}", e);
-                None
-            }
-        }
-    }
-
-    fn add_surface_to_focused_pane(&mut self) -> Option<u64> {
-        let sf_id = self.next_surface_id;
-        self.next_surface_id += 1;
-        let ws_id = self.active_workspace().id;
-        let focused = self.focused_pane_id();
-
-        match spawn_surface(
-            80,
-            24,
-            &self.socket_addr,
-            &self.socket_token,
-            &self.config,
-            ws_id,
-            sf_id,
-            None,
-            None,
-        ) {
-            Ok(surface) => {
-                if let Some(managed) = self.panes.get_mut(&focused) {
-                    managed.surfaces.push(surface);
-                    managed.active_surface_idx = managed.surfaces.len() - 1;
-                    Some(sf_id)
-                } else {
-                    None
-                }
-            }
-            Err(e) => {
-                tracing::error!("Failed to spawn surface: {}", e);
-                None
-            }
-        }
-    }
-
-    fn close_workspace_at(&mut self, ws_idx: usize) {
-        let ws_id = self.workspaces[ws_idx].id;
-        if self.workspaces.len() <= 1 {
-            // Last workspace — clean up and signal exit
-            let pane_ids: Vec<PaneId> = self.workspaces[ws_idx].tree.iter_panes();
-            for id in &pane_ids {
-                self.panes.remove(id);
-                self.notifications.remove_pane(*id);
-            }
-            self.notifications.remove_workspace(ws_id);
-            self.wants_exit = true;
-            return;
-        }
-        let pane_ids: Vec<PaneId> = self.workspaces[ws_idx].tree.iter_panes();
-        for id in &pane_ids {
-            self.panes.remove(id);
-            self.notifications.remove_pane(*id);
-        }
-        self.notifications.remove_workspace(ws_id);
-        self.workspaces.remove(ws_idx);
-        if self.active_workspace_idx >= self.workspaces.len() {
-            self.active_workspace_idx = self.workspaces.len() - 1;
-        }
-    }
-
-    // --- Menu bar actions ---
-
-    fn handle_menu_actions(&mut self) {
-        while let Some(action) = menu_bar::take_pending_action() {
-            match action {
-                menu_bar::MenuAction::NewWorkspace => {
-                    self.create_workspace(None);
-                }
-                menu_bar::MenuAction::NewTab => {
-                    self.add_surface_to_focused_pane();
-                }
-                menu_bar::MenuAction::CloseTab => {
-                    self.do_close_cascade();
-                }
-                menu_bar::MenuAction::SaveSession => {
-                    self.flush_pending_io();
-                    let data = self.build_session_data();
-                    if let Err(e) = amux_session::save(&data) {
-                        tracing::error!("Failed to save session: {}", e);
-                    }
-                }
-                menu_bar::MenuAction::ToggleSidebar => {
-                    self.sidebar.visible = !self.sidebar.visible;
-                }
-                menu_bar::MenuAction::ToggleNotificationPanel => {
-                    self.show_notification_panel = !self.show_notification_panel;
-                }
-                menu_bar::MenuAction::ZoomIn => {
-                    self.font_size = (self.font_size + 1.0).min(96.0);
-                    #[cfg(feature = "gpu-renderer")]
-                    if let Some(gpu) = &mut self.gpu_renderer {
-                        gpu.set_font_size(self.font_size);
-                    }
-                }
-                menu_bar::MenuAction::ZoomOut => {
-                    self.font_size = (self.font_size - 1.0).max(4.0);
-                    #[cfg(feature = "gpu-renderer")]
-                    if let Some(gpu) = &mut self.gpu_renderer {
-                        gpu.set_font_size(self.font_size);
-                    }
-                }
-                menu_bar::MenuAction::ZoomReset => {
-                    self.font_size = font::DEFAULT_FONT_SIZE;
-                    #[cfg(feature = "gpu-renderer")]
-                    if let Some(gpu) = &mut self.gpu_renderer {
-                        gpu.set_font_size(self.font_size);
-                    }
-                }
-            }
-        }
-    }
-
-    // --- Shortcuts ---
-
-    fn handle_shortcuts(&mut self, ctx: &egui::Context) -> bool {
-        // Skip terminal shortcuts when a modal text field has focus — let egui
-        // handle Cmd+V, Cmd+C, etc. for the text widget instead.
-        if self.rename_modal.is_some() || self.find_state.is_some() {
-            return false;
-        }
-        let events = ctx.input(|i| i.events.clone());
-
-        for event in &events {
-            if let egui::Event::Key {
-                key,
-                pressed: true,
-                modifiers,
-                ..
-            } = event
-            {
-                #[cfg(target_os = "macos")]
-                let is_cmd = modifiers.mac_cmd || modifiers.command;
-                #[cfg(not(target_os = "macos"))]
-                let is_cmd = modifiers.ctrl && modifiers.shift;
-
-                // Copy: Cmd+C (with selection) / Cmd+Shift+C (always copy)
-                #[cfg(target_os = "macos")]
-                let is_copy = is_cmd && (*key == egui::Key::C);
-                #[cfg(not(target_os = "macos"))]
-                let is_copy = modifiers.ctrl && modifiers.shift && (*key == egui::Key::C);
-
-                // Copy selection if active; otherwise fall through to send Ctrl+C
-                if is_copy && self.copy_selection() {
-                    return true;
-                }
-
-                // Paste: Cmd+V (macOS) / Ctrl+Shift+V (other)
-                #[cfg(target_os = "macos")]
-                let is_paste = is_cmd && !modifiers.shift && *key == egui::Key::V;
-                #[cfg(not(target_os = "macos"))]
-                let is_paste = modifiers.ctrl && modifiers.shift && *key == egui::Key::V;
-
-                if is_paste {
-                    if let Ok(mut clipboard) = arboard::Clipboard::new() {
-                        if let Ok(text) = clipboard.get_text() {
-                            if !text.is_empty() {
-                                self.do_paste(&text);
-                            }
-                        }
-                    }
-                    return true;
-                }
-
-                // Copy mode: intercept all keys when active
-                if self.copy_mode.is_some() {
-                    return self.handle_copy_mode_key(key, modifiers);
-                }
-
-                // Escape: close find bar, exit copy mode, or clear selection
-                if *key == egui::Key::Escape
-                    && !modifiers.shift
-                    && !modifiers.ctrl
-                    && !modifiers.alt
-                {
-                    if self.find_state.is_some() {
-                        self.find_state = None;
-                        return true;
-                    }
-                    let focused = self.focused_pane_id();
-                    if let Some(m) = self.panes.get_mut(&focused) {
-                        if m.selection.is_some() {
-                            m.selection = None;
-                            return true;
-                        }
-                    }
-                }
-
-                // Find: Cmd+F (macOS) / Ctrl+Shift+F (other)
-                if is_cmd && !modifiers.shift && *key == egui::Key::F {
-                    let pane_id = self.focused_pane_id();
-                    self.find_state = Some(FindState {
-                        query: String::new(),
-                        matches: Vec::new(),
-                        current_match: 0,
-                        pane_id,
-                        just_opened: true,
-                    });
-                    return true;
-                }
-
-                // Select all: Cmd+A
-                #[cfg(target_os = "macos")]
-                if is_cmd && !modifiers.shift && *key == egui::Key::A {
-                    self.select_all_visible();
-                    return true;
-                }
-                #[cfg(not(target_os = "macos"))]
-                if modifiers.ctrl && modifiers.shift && *key == egui::Key::A {
-                    self.select_all_visible();
-                    return true;
-                }
-
-                // Enter copy mode: Cmd+Shift+X (macOS) / Ctrl+Shift+X (other)
-                if is_cmd && modifiers.shift && *key == egui::Key::X {
-                    self.enter_copy_mode();
-                    return true;
-                }
-
-                // Toggle sidebar: Cmd+B / Ctrl+B
-                #[cfg(target_os = "macos")]
-                if is_cmd && !modifiers.shift && *key == egui::Key::B {
-                    self.sidebar.visible = !self.sidebar.visible;
-                    return true;
-                }
-                #[cfg(not(target_os = "macos"))]
-                if modifiers.ctrl && !modifiers.shift && *key == egui::Key::B {
-                    self.sidebar.visible = !self.sidebar.visible;
-                    return true;
-                }
-
-                // New workspace: Cmd+N / Ctrl+Shift+N
-                #[cfg(target_os = "macos")]
-                if is_cmd && !modifiers.shift && *key == egui::Key::N {
-                    self.create_workspace(None);
-                    return true;
-                }
-                #[cfg(not(target_os = "macos"))]
-                if modifiers.ctrl && modifiers.shift && *key == egui::Key::N {
-                    self.create_workspace(None);
-                    return true;
-                }
-
-                // New tab in focused pane: Cmd+T / Ctrl+Shift+T
-                #[cfg(target_os = "macos")]
-                if is_cmd && !modifiers.shift && *key == egui::Key::T {
-                    self.add_surface_to_focused_pane();
-                    return true;
-                }
-                #[cfg(not(target_os = "macos"))]
-                if modifiers.ctrl && modifiers.shift && *key == egui::Key::T {
-                    self.add_surface_to_focused_pane();
-                    return true;
-                }
-
-                // Next workspace: Cmd+Shift+]
-                #[cfg(target_os = "macos")]
-                if is_cmd && modifiers.shift && *key == egui::Key::CloseBracket {
-                    if self.workspaces.len() > 1 {
-                        self.active_workspace_idx =
-                            (self.active_workspace_idx + 1) % self.workspaces.len();
-                    }
-                    return true;
-                }
-                #[cfg(not(target_os = "macos"))]
-                if modifiers.ctrl && modifiers.shift && *key == egui::Key::CloseBracket {
-                    if self.workspaces.len() > 1 {
-                        self.active_workspace_idx =
-                            (self.active_workspace_idx + 1) % self.workspaces.len();
-                    }
-                    return true;
-                }
-
-                // Prev workspace: Cmd+Shift+[
-                #[cfg(target_os = "macos")]
-                if is_cmd && modifiers.shift && *key == egui::Key::OpenBracket {
-                    if self.workspaces.len() > 1 {
-                        self.active_workspace_idx = if self.active_workspace_idx == 0 {
-                            self.workspaces.len() - 1
-                        } else {
-                            self.active_workspace_idx - 1
-                        };
-                    }
-                    return true;
-                }
-                #[cfg(not(target_os = "macos"))]
-                if modifiers.ctrl && modifiers.shift && *key == egui::Key::OpenBracket {
-                    if self.workspaces.len() > 1 {
-                        self.active_workspace_idx = if self.active_workspace_idx == 0 {
-                            self.workspaces.len() - 1
-                        } else {
-                            self.active_workspace_idx - 1
-                        };
-                    }
-                    return true;
-                }
-
-                // Jump to workspace 1-9 (Cmd+9 = last workspace)
-                #[cfg(target_os = "macos")]
-                let is_jump_mod = is_cmd && !modifiers.shift;
-                #[cfg(not(target_os = "macos"))]
-                let is_jump_mod = modifiers.ctrl && !modifiers.shift;
-
-                if is_jump_mod {
-                    let num = match key {
-                        egui::Key::Num1 => Some(0usize),
-                        egui::Key::Num2 => Some(1),
-                        egui::Key::Num3 => Some(2),
-                        egui::Key::Num4 => Some(3),
-                        egui::Key::Num5 => Some(4),
-                        egui::Key::Num6 => Some(5),
-                        egui::Key::Num7 => Some(6),
-                        egui::Key::Num8 => Some(7),
-                        egui::Key::Num9 => Some(usize::MAX), // last workspace
-                        _ => None,
-                    };
-                    if let Some(mut idx) = num {
-                        if idx == usize::MAX {
-                            idx = self.workspaces.len().saturating_sub(1);
-                        }
-                        if idx < self.workspaces.len() {
-                            self.active_workspace_idx = idx;
-                            return true;
-                        }
-                    }
-                }
-
-                // Next tab in focused pane: Ctrl+Tab
-                if modifiers.ctrl && !modifiers.shift && *key == egui::Key::Tab {
-                    if let Some(managed) = self.panes.get_mut(&self.focused_pane_id()) {
-                        if managed.surfaces.len() > 1 {
-                            managed.active_surface_idx =
-                                (managed.active_surface_idx + 1) % managed.surfaces.len();
-                        }
-                    }
-                    return true;
-                }
-
-                // Prev tab in focused pane: Ctrl+Shift+Tab
-                if modifiers.ctrl && modifiers.shift && *key == egui::Key::Tab {
-                    if let Some(managed) = self.panes.get_mut(&self.focused_pane_id()) {
-                        if managed.surfaces.len() > 1 {
-                            managed.active_surface_idx = if managed.active_surface_idx == 0 {
-                                managed.surfaces.len() - 1
-                            } else {
-                                managed.active_surface_idx - 1
-                            };
-                        }
-                    }
-                    return true;
-                }
-
-                // --- Pane shortcuts ---
-
-                // Split right: Cmd+D (macOS) / Ctrl+Shift+D (other)
-                #[cfg(target_os = "macos")]
-                if is_cmd && !modifiers.shift && *key == egui::Key::D {
-                    return self.do_split(SplitDirection::Horizontal);
-                }
-                #[cfg(not(target_os = "macos"))]
-                if is_cmd && *key == egui::Key::D {
-                    return self.do_split(SplitDirection::Horizontal);
-                }
-                // Split down: Cmd+Shift+D (macOS) / Ctrl+Shift+Down (other)
-                #[cfg(target_os = "macos")]
-                if is_cmd && modifiers.shift && *key == egui::Key::D {
-                    return self.do_split(SplitDirection::Vertical);
-                }
-                #[cfg(not(target_os = "macos"))]
-                if modifiers.ctrl && modifiers.shift && *key == egui::Key::ArrowDown {
-                    return self.do_split(SplitDirection::Vertical);
-                }
-
-                // Close: Cmd+W — cascade: tab -> pane -> workspace
-                if is_cmd && *key == egui::Key::W {
-                    return self.do_close_cascade();
-                }
-
-                // Navigate: Option+Cmd+Arrow / Ctrl+Alt+Arrow
-                #[cfg(target_os = "macos")]
-                let is_nav = is_cmd && modifiers.alt;
-                #[cfg(not(target_os = "macos"))]
-                let is_nav = modifiers.ctrl && modifiers.alt;
-
-                if is_nav {
-                    let dir = match key {
-                        egui::Key::ArrowLeft => Some(NavDirection::Left),
-                        egui::Key::ArrowRight => Some(NavDirection::Right),
-                        egui::Key::ArrowUp => Some(NavDirection::Up),
-                        egui::Key::ArrowDown => Some(NavDirection::Down),
-                        _ => None,
-                    };
-                    if let Some(dir) = dir {
-                        return self.do_navigate(dir);
-                    }
-                }
-
-                // Zoom toggle: Cmd+Shift+Enter / Ctrl+Shift+Enter
-                #[cfg(target_os = "macos")]
-                let is_zoom = is_cmd && modifiers.shift && *key == egui::Key::Enter;
-                #[cfg(not(target_os = "macos"))]
-                let is_zoom = modifiers.ctrl && modifiers.shift && *key == egui::Key::Enter;
-
-                if is_zoom {
-                    return self.do_toggle_zoom();
-                }
-
-                // Notification panel: Cmd+I / Ctrl+I
-                if is_cmd && !modifiers.shift && *key == egui::Key::I {
-                    self.show_notification_panel = !self.show_notification_panel;
-                    return true;
-                }
-
-                // Jump to latest unread: Cmd+Shift+U / Ctrl+Shift+U
-                if is_cmd && modifiers.shift && *key == egui::Key::U {
-                    self.jump_to_latest_unread();
-                    return true;
-                }
-
-                // Clear scrollback: Cmd+K (macOS) / Ctrl+Shift+K (other)
-                if is_cmd && !modifiers.shift && *key == egui::Key::K {
-                    self.do_clear_scrollback();
-                    return true;
-                }
-
-                // Scroll
-                if modifiers.shift && *key == egui::Key::PageUp {
-                    return self.do_scroll(-1);
-                }
-                if modifiers.shift && *key == egui::Key::PageDown {
-                    return self.do_scroll(1);
-                }
-            }
-        }
-
-        // Mouse wheel scrolling — scroll the pane under the cursor
-        let scroll_delta = ctx.input(|i| i.smooth_scroll_delta.y);
-        if scroll_delta != 0.0 {
-            let hover_pos = ctx.input(|i| i.pointer.hover_pos());
-            let target_pane = hover_pos.and_then(|pos| {
-                let panel_rect = self.last_panel_rect?;
-                let ws = self.active_workspace();
-                if let Some(zoomed_id) = ws.zoomed {
-                    // In zoomed mode, only the zoomed pane is visible
-                    if panel_rect.contains(pos) {
-                        return Some(zoomed_id);
-                    }
-                    return None;
-                }
-                let layout = ws.tree.layout(panel_rect);
-                layout
-                    .iter()
-                    .find(|(_, rect)| rect.contains(pos))
-                    .map(|(id, _)| *id)
-            });
-            if let Some(pane_id) = target_pane {
-                if let Some(managed) = self.panes.get_mut(&pane_id) {
-                    let surface = managed.active_surface_mut();
-                    let font_id = egui::FontId::monospace(self.font_size);
-                    let cell_height = ctx.fonts(|f| f.row_height(&font_id));
-
-                    surface.scroll_accum += -scroll_delta / cell_height;
-                    let whole_lines = surface.scroll_accum.trunc() as isize;
-                    if whole_lines != 0 {
-                        surface.scroll_accum -= whole_lines as f32;
-                        self.do_scroll_lines_for(pane_id, whole_lines);
-                    }
-                }
-            }
-        }
-
-        false
-    }
-
-    fn do_split(&mut self, direction: SplitDirection) -> bool {
-        let Some(new_id) = self.spawn_pane_with_surface() else {
-            return false;
-        };
-        let ws = self.active_workspace_mut();
-        if ws.tree.split(ws.focused_pane, direction, new_id) {
-            self.set_focus(new_id);
-            true
-        } else {
-            // Split failed — clean up the spawned pane
-            self.panes.remove(&new_id);
-            false
-        }
-    }
-
-    fn do_close_cascade(&mut self) -> bool {
-        let focused_id = self.focused_pane_id();
-
-        // First check: close a tab if >1 tab in focused pane
-        if let Some(managed) = self.panes.get_mut(&focused_id) {
-            if managed.surfaces.len() > 1 {
-                managed.surfaces.remove(managed.active_surface_idx);
-                if managed.active_surface_idx >= managed.surfaces.len() {
-                    managed.active_surface_idx = managed.surfaces.len() - 1;
-                }
-                return true;
-            }
-        }
-
-        self.close_pane(focused_id);
-        true
-    }
-
-    /// Restart the active surface in a pane by spawning a new shell in the
-    /// same working directory and replacing the dead surface in-place.
-    fn restart_surface(&mut self, pane_id: PaneId) {
-        let ws_id = self
-            .workspaces
-            .iter()
-            .find(|ws| ws.tree.iter_panes().contains(&pane_id))
-            .map(|ws| ws.id)
-            .unwrap_or(0);
-
-        let managed = match self.panes.get_mut(&pane_id) {
-            Some(m) => m,
-            None => return,
-        };
-        let old_surface = managed.active_surface_mut();
-        let cwd = old_surface.metadata.cwd.clone();
-        let (cols, rows) = old_surface.pane.dimensions();
-        let sf_id = self.next_surface_id;
-        self.next_surface_id += 1;
-
-        match spawn_surface(
-            cols as u16,
-            rows as u16,
-            &self.socket_addr,
-            &self.socket_token,
-            &self.config,
-            ws_id,
-            sf_id,
-            cwd.as_deref(),
-            None,
-        ) {
-            Ok(new_surface) => {
-                let idx = managed.active_surface_idx;
-                managed.surfaces[idx] = new_surface;
-            }
-            Err(e) => {
-                tracing::warn!("Failed to restart surface: {e}");
-            }
-        }
-    }
-
-    /// Close a pane entirely. Finds the owning workspace (not necessarily the
-    /// active one). If it's the last pane in that workspace, close the workspace.
-    fn close_pane(&mut self, pane_id: PaneId) {
-        // Find the workspace that owns this pane
-        let ws_idx = match self
-            .workspaces
-            .iter()
-            .position(|ws| ws.tree.iter_panes().contains(&pane_id))
-        {
-            Some(idx) => idx,
-            None => return, // pane not in any workspace
-        };
-
-        let pane_count = self.workspaces[ws_idx].tree.iter_panes().len();
-        if pane_count > 1 {
-            let ws = &mut self.workspaces[ws_idx];
-            if let Some(new_focus) = ws.tree.close(pane_id) {
-                ws.last_pane_sizes.remove(&pane_id);
-                if ws.zoomed == Some(pane_id) {
-                    ws.zoomed = None;
-                }
-                self.panes.remove(&pane_id);
-                self.notifications.remove_pane(pane_id);
-                if ws_idx == self.active_workspace_idx {
-                    self.set_focus(new_focus);
-                }
-            }
-        } else {
-            // Last pane in workspace -> close workspace
-            self.close_workspace_at(ws_idx);
-        }
-    }
-
-    fn do_navigate(&mut self, dir: NavDirection) -> bool {
-        if let Some(rect) = self.last_panel_rect {
-            let ws = self.active_workspace();
-            if let Some(neighbor) = ws.tree.neighbor(ws.focused_pane, dir, rect) {
-                self.set_focus(neighbor);
-            } else {
-                self.flash_focus();
-            }
-        }
-        true
-    }
-
-    fn do_toggle_zoom(&mut self) -> bool {
-        let ws = self.active_workspace_mut();
-        if ws.zoomed.is_some() {
-            ws.zoomed = None;
-        } else {
-            ws.zoomed = Some(ws.focused_pane);
-        }
-        true
-    }
-
-    fn do_scroll(&mut self, pages: isize) -> bool {
-        let focused_id = self.focused_pane_id();
-        if let Some(managed) = self.panes.get_mut(&focused_id) {
-            let surface = managed.active_surface_mut();
-            let (_, rows) = surface.pane.dimensions();
-            let page_size = rows.saturating_sub(1).max(1);
-            let lines = pages * page_size as isize;
-            let total = surface.pane.screen().scrollback_rows();
-            let max_offset = total.saturating_sub(rows);
-            let new_offset = surface.scroll_offset as isize - lines;
-            surface.scroll_offset = (new_offset.max(0) as usize).min(max_offset);
-        }
-        true
-    }
-
-    fn do_scroll_lines_for(&mut self, pane_id: PaneId, lines: isize) {
-        if let Some(managed) = self.panes.get_mut(&pane_id) {
-            let surface = managed.active_surface_mut();
-            let (_, rows) = surface.pane.dimensions();
-            let total = surface.pane.screen().scrollback_rows();
-            let max_offset = total.saturating_sub(rows);
-            let new_offset = surface.scroll_offset as isize - lines;
-            surface.scroll_offset = (new_offset.max(0) as usize).min(max_offset);
-        }
-    }
-
-    fn do_clear_scrollback(&mut self) {
-        let focused_id = self.focused_pane_id();
-        if let Some(managed) = self.panes.get_mut(&focused_id) {
-            let surface = managed.active_surface_mut();
-            // 1. Clear visible screen and move cursor home via terminal state machine
-            surface.pane.feed_bytes(b"\x1b[2J\x1b[H");
-            // 2. Erase scrollback buffer
-            surface.pane.erase_scrollback();
-            surface.scroll_offset = 0;
-            surface.scroll_accum = 0.0;
-            // 3. Send Ctrl+L to the PTY so the shell redraws its prompt
-            let _ = surface.pane.write_bytes(b"\x0c");
-        }
-    }
-
-    // --- Notifications ---
-
-    fn drain_notifications(&mut self) {
-        // Collect events first to avoid borrow conflicts
-        let mut events: Vec<(u64, u64, u64, NotificationEvent)> = Vec::new();
-
-        for (&pane_id, managed) in &self.panes {
-            let ws_id = self.workspace_for_pane(pane_id).unwrap_or(0);
-            for surface in &managed.surfaces {
-                for event in surface.pane.drain_notifications() {
-                    events.push((ws_id, pane_id, surface.id, event));
-                }
-            }
-        }
-
-        for (ws_id, pane_id, surface_id, event) in events {
-            let (title, body, source) = match event {
-                NotificationEvent::Toast { title, body } => {
-                    (title.unwrap_or_default(), body, NotificationSource::Toast)
-                }
-                NotificationEvent::Bell => {
-                    (String::new(), "Bell".to_string(), NotificationSource::Bell)
-                }
-                NotificationEvent::TitleChanged(_) => {
-                    continue;
-                }
-                NotificationEvent::WorkingDirectoryChanged => {
-                    // Store the CWD from OSC 7 into surface metadata
-                    if let Some(managed) = self.panes.get_mut(&pane_id) {
-                        for surface in &mut managed.surfaces {
-                            if surface.id == surface_id {
-                                let cwd = surface
-                                    .pane
-                                    .working_dir()
-                                    .and_then(|url| url.to_file_path().ok())
-                                    .map(|p| p.to_string_lossy().to_string());
-                                if cwd.is_some() {
-                                    surface.metadata.cwd = cwd;
-                                }
-                            }
-                        }
-                    }
-                    continue;
-                }
-            };
-
-            let skip_toast = matches!(source, NotificationSource::Bell);
-            self.deliver_notification(
-                ws_id,
-                pane_id,
-                surface_id,
-                title,
-                String::new(),
-                body,
-                source,
-                skip_toast,
-            );
-        }
-    }
-
-    /// Three-tier notification delivery (matching cmux):
-    /// 1. App unfocused → system toast + custom command + unread
-    /// 2. App focused, different pane → in-app sound + custom command + unread
-    /// 3. App focused, same pane → mark read (flash only, no ring)
-    ///
-    /// `skip_toast` suppresses the system toast (used for bell notifications).
-    #[allow(clippy::too_many_arguments)]
-    fn deliver_notification(
-        &mut self,
-        ws_id: u64,
-        pane_id: PaneId,
-        surface_id: u64,
-        title: String,
-        subtitle: String,
-        body: String,
-        source: NotificationSource,
-        skip_toast: bool,
-    ) -> u64 {
-        let focused = self.focused_pane_id();
-        let source_str = source.as_str();
-        // Clone for the IPC broadcast after the notification is stored.
-        let bc_title = title.clone();
-        let bc_subtitle = subtitle.clone();
-        let bc_body = body.clone();
-
-        let nid = if !self.app_focused {
-            // Tier 1: app is unfocused — always treat as background
-            if self.app_config.notifications.system_notifications && !skip_toast {
-                self.system_notifier.send(&title, &body, ws_id, pane_id);
-            }
-            if let Some(cmd) = &self.app_config.notifications.custom_command {
-                self.system_notifier
-                    .run_custom_command(cmd, &title, &body, source_str);
-            }
-            let nid = self
-                .notifications
-                .push(ws_id, pane_id, surface_id, title, subtitle, body, source);
-            if self.app_config.notifications.auto_reorder_workspaces {
-                self.bubble_workspace(ws_id);
-            }
-            nid
-        } else if pane_id == focused {
-            // Tier 3: app focused, same pane — mark read (flash only)
-            self.notifications
-                .push_read(ws_id, pane_id, surface_id, title, subtitle, body, source)
-        } else {
-            // Tier 2: app focused, different pane — in-app sound + command
-            if self.app_config.notifications.sound.play_when_focused {
-                if let Some(player) = &self.sound_player {
-                    player.play();
-                }
-            }
-            if let Some(cmd) = &self.app_config.notifications.custom_command {
-                self.system_notifier
-                    .run_custom_command(cmd, &title, &body, source_str);
-            }
-            let nid = self
-                .notifications
-                .push(ws_id, pane_id, surface_id, title, subtitle, body, source);
-            if self.app_config.notifications.auto_reorder_workspaces {
-                self.bubble_workspace(ws_id);
-            }
-            nid
-        };
-
-        // Broadcast to subscribed IPC clients
-        self.event_broadcaster.send(amux_ipc::ServerEvent {
-            event: "notification".to_string(),
-            data: serde_json::json!({
-                "notification_id": nid,
-                "workspace_id": ws_id.to_string(),
-                "pane_id": pane_id.to_string(),
-                "title": bc_title,
-                "subtitle": bc_subtitle,
-                "body": bc_body,
-                "source": source_str,
-            }),
-        });
-
-        nid
-    }
-
-    /// Move a workspace to the top of the sidebar (just index 0 for now,
-    /// since amux doesn't have pinned workspaces yet). Adjusts
-    /// `active_workspace_idx` to keep the active workspace correct.
-    fn bubble_workspace(&mut self, workspace_id: u64) {
-        let active_ws_id = self.workspaces[self.active_workspace_idx].id;
-        // Don't bubble the active workspace
-        if workspace_id == active_ws_id {
-            return;
-        }
-        let Some(from) = self.workspaces.iter().position(|ws| ws.id == workspace_id) else {
-            return;
-        };
-        if from == 0 {
-            return;
-        }
-        let ws = self.workspaces.remove(from);
-        self.workspaces.insert(0, ws);
-        // Fix active_workspace_idx: the active workspace shifted right by 1
-        // if it was before the removed position.
-        self.active_workspace_idx = self
-            .workspaces
-            .iter()
-            .position(|ws| ws.id == active_ws_id)
-            .unwrap_or(0);
-    }
-
-    fn workspace_for_pane(&self, pane_id: PaneId) -> Option<u64> {
-        self.workspaces
-            .iter()
-            .find(|ws| ws.tree.iter_panes().contains(&pane_id))
-            .map(|ws| ws.id)
-    }
-
-    /// Aggregate metadata from the focused surface of the focused pane in a workspace.
-    fn workspace_metadata(&self, workspace: &Workspace) -> SurfaceMetadata {
-        self.panes
-            .get(&workspace.focused_pane)
-            .map(|mp| {
-                let sf = mp.active_surface();
-                let mut meta = sf.metadata.clone();
-                // Capture the surface's OSC title for sidebar display
-                let title = sf.pane.title();
-                if !title.is_empty() {
-                    meta.surface_title = Some(title.to_string());
-                }
-                meta
-            })
-            .unwrap_or_default()
-    }
-
-    fn jump_to_latest_unread(&mut self) {
-        if let Some(notif) = self.notifications.most_recent_unread() {
-            let ws_id = notif.workspace_id;
-            let pane_id = notif.pane_id as PaneId;
-
-            // Switch to the notification's workspace
-            if let Some(idx) = self.workspaces.iter().position(|ws| ws.id == ws_id) {
-                self.active_workspace_idx = idx;
-            }
-            self.set_focus(pane_id);
-        }
-    }
-
     fn update_ime_position(&self, ctx: &egui::Context) {
         let focused_id = self.focused_pane_id();
         let panel_rect = match self.last_panel_rect {
@@ -3153,197 +1503,6 @@ impl AmuxApp {
         }
     }
 
-    fn enter_copy_mode(&mut self) {
-        let pane_id = self.focused_pane_id();
-        if let Some(managed) = self.panes.get(&pane_id) {
-            let surface = managed.active_surface();
-            let cursor = surface.pane.cursor();
-            let screen = surface.pane.screen();
-            let (_, rows) = surface.pane.dimensions();
-            let total = screen.scrollback_rows();
-            let end = total.saturating_sub(surface.scroll_offset);
-            let start = end.saturating_sub(rows);
-            // Place copy mode cursor at terminal cursor position in phys coords
-            let phys_row = start + (cursor.y.max(0) as usize).min(rows.saturating_sub(1));
-            self.copy_mode = Some(CopyModeState {
-                pane_id,
-                cursor: (cursor.x, phys_row),
-                visual_anchor: None,
-                line_visual: false,
-            });
-        }
-    }
-
-    fn handle_copy_mode_key(&mut self, key: &egui::Key, modifiers: &egui::Modifiers) -> bool {
-        let cm = match self.copy_mode.as_mut() {
-            Some(cm) => cm,
-            None => return false,
-        };
-        let pane_id = cm.pane_id;
-
-        // Get dimensions for bounds checking
-        let (cols, rows, total_rows) = match self.panes.get(&pane_id) {
-            Some(m) => {
-                let s = m.active_surface();
-                let (c, r) = s.pane.dimensions();
-                let t = s.pane.screen().scrollback_rows();
-                (c, r, t)
-            }
-            None => {
-                self.copy_mode = None;
-                return true;
-            }
-        };
-        let cm = self.copy_mode.as_mut().unwrap();
-
-        match key {
-            // Exit copy mode
-            egui::Key::Escape | egui::Key::Q => {
-                self.copy_mode = None;
-                return true;
-            }
-            // Movement
-            egui::Key::H | egui::Key::ArrowLeft => {
-                cm.cursor.0 = cm.cursor.0.saturating_sub(1);
-            }
-            egui::Key::L | egui::Key::ArrowRight => {
-                cm.cursor.0 = (cm.cursor.0 + 1).min(cols.saturating_sub(1));
-            }
-            egui::Key::K | egui::Key::ArrowUp => {
-                cm.cursor.1 = cm.cursor.1.saturating_sub(1);
-            }
-            egui::Key::J | egui::Key::ArrowDown => {
-                cm.cursor.1 = (cm.cursor.1 + 1).min(total_rows.saturating_sub(1));
-            }
-            // Half-page up/down
-            egui::Key::U if modifiers.ctrl => {
-                let half = rows / 2;
-                cm.cursor.1 = cm.cursor.1.saturating_sub(half);
-            }
-            egui::Key::D if modifiers.ctrl => {
-                let half = rows / 2;
-                cm.cursor.1 = (cm.cursor.1 + half).min(total_rows.saturating_sub(1));
-            }
-            // End of scrollback (Shift+G = vim 'G')
-            egui::Key::G if modifiers.shift => {
-                cm.cursor.1 = total_rows.saturating_sub(1);
-                cm.cursor.0 = 0;
-            }
-            // Start of scrollback (g = vim 'gg', second g handled by repeat)
-            egui::Key::G => {
-                cm.cursor.1 = 0;
-                cm.cursor.0 = 0;
-            }
-            // Line start/end
-            egui::Key::Num0 => {
-                cm.cursor.0 = 0;
-            }
-            // Visual mode toggle
-            egui::Key::V if modifiers.shift => {
-                // Line visual
-                if cm.line_visual {
-                    cm.visual_anchor = None;
-                    cm.line_visual = false;
-                } else {
-                    cm.visual_anchor = Some(cm.cursor);
-                    cm.line_visual = true;
-                }
-            }
-            egui::Key::V => {
-                // Character visual
-                if cm.visual_anchor.is_some() && !cm.line_visual {
-                    cm.visual_anchor = None;
-                } else {
-                    cm.visual_anchor = Some(cm.cursor);
-                    cm.line_visual = false;
-                }
-            }
-            // Yank selection
-            egui::Key::Y => {
-                let anchor = cm.visual_anchor;
-                let line_visual = cm.line_visual;
-                if let Some(anchor) = anchor {
-                    let text = self.extract_copy_mode_text(pane_id, anchor, cols, line_visual);
-                    if let Some(text) = text {
-                        if let Ok(mut clipboard) = arboard::Clipboard::new() {
-                            let _ = clipboard.set_text(&text);
-                        }
-                    }
-                    self.copy_mode = None;
-                    return true;
-                }
-            }
-            _ => {}
-        }
-
-        // Scroll viewport to keep cursor visible
-        if let Some(managed) = self.panes.get_mut(&pane_id) {
-            let cm = self.copy_mode.as_ref().unwrap();
-            let surface = managed.active_surface_mut();
-            let end = total_rows.saturating_sub(surface.scroll_offset);
-            let start = end.saturating_sub(rows);
-            if cm.cursor.1 < start {
-                surface.scroll_offset = total_rows.saturating_sub(cm.cursor.1 + rows);
-            } else if cm.cursor.1 >= end {
-                surface.scroll_offset = total_rows.saturating_sub(cm.cursor.1 + 1);
-            }
-        }
-
-        true
-    }
-
-    fn extract_copy_mode_text(
-        &self,
-        pane_id: PaneId,
-        anchor: (usize, usize),
-        cols: usize,
-        line_visual: bool,
-    ) -> Option<String> {
-        let cm = self.copy_mode.as_ref()?;
-        let managed = self.panes.get(&pane_id)?;
-        let surface = managed.active_surface();
-        let screen = surface.pane.screen();
-
-        let (start, end) =
-            if anchor.1 < cm.cursor.1 || (anchor.1 == cm.cursor.1 && anchor.0 <= cm.cursor.0) {
-                (anchor, cm.cursor)
-            } else {
-                (cm.cursor, anchor)
-            };
-
-        let lines = screen.lines_in_phys_range(start.1..end.1 + 1);
-        let mut result = String::new();
-
-        for (i, line) in lines.iter().enumerate() {
-            if i > 0 {
-                result.push('\n');
-            }
-            let phys_row = start.1 + i;
-            for cell in line.visible_cells() {
-                let col = cell.cell_index();
-                if col >= cols {
-                    break;
-                }
-                if line_visual {
-                    result.push_str(cell.str());
-                } else {
-                    // Character visual: clip to selection bounds
-                    if phys_row == start.1 && col < start.0 {
-                        continue;
-                    }
-                    if phys_row == end.1 && col > end.0 {
-                        break;
-                    }
-                    result.push_str(cell.str());
-                }
-            }
-        }
-
-        // Trim trailing whitespace per line
-        let trimmed: Vec<&str> = result.lines().map(|l| l.trim_end()).collect();
-        Some(trimmed.join("\n"))
-    }
-
     fn render_rename_modal(&mut self, ctx: &egui::Context) {
         let mut apply: Option<String> = None;
         let mut cancel = false;
@@ -3526,162 +1685,6 @@ impl AmuxApp {
         }
     }
 
-    fn render_notification_panel(&mut self, ctx: &egui::Context) {
-        let mut close_panel = false;
-        let mut mark_all = false;
-        let mut jump_to: Option<(u64, u64)> = None; // (workspace_id, pane_id)
-        let mut remove_id: Option<u64> = None;
-
-        egui::Window::new("Notifications")
-            .collapsible(false)
-            .resizable(true)
-            .default_size([380.0, 460.0])
-            .anchor(egui::Align2::RIGHT_TOP, [-10.0, 10.0])
-            .show(ctx, |ui| {
-                ui.horizontal(|ui| {
-                    ui.heading(
-                        egui::RichText::new("Notifications").color(egui::Color32::from_gray(220)),
-                    );
-                    ui.with_layout(egui::Layout::right_to_left(egui::Align::Center), |ui| {
-                        if ui.small_button("Clear All").clicked() {
-                            mark_all = true;
-                        }
-                        if ui.small_button("Jump to Latest").clicked() {
-                            jump_to = self
-                                .notifications
-                                .most_recent_unread()
-                                .map(|n| (n.workspace_id, n.pane_id));
-                            close_panel = true;
-                        }
-                    });
-                });
-                ui.separator();
-
-                let notifications = self.notifications.all_notifications();
-                if notifications.is_empty() {
-                    ui.vertical_centered(|ui| {
-                        ui.add_space(40.0);
-                        ui.label(
-                            egui::RichText::new("\u{1f515}") // 🔕
-                                .size(32.0),
-                        );
-                        ui.add_space(8.0);
-                        ui.label(
-                            egui::RichText::new("No notifications")
-                                .color(egui::Color32::from_gray(140))
-                                .size(14.0),
-                        );
-                        ui.label(
-                            egui::RichText::new("Agent notifications will appear here")
-                                .color(egui::Color32::from_gray(80))
-                                .size(11.0),
-                        );
-                    });
-                } else {
-                    egui::ScrollArea::vertical().show(ui, |ui| {
-                        // Group by workspace, most recent notifications first within each
-                        let mut grouped: Vec<_> = notifications.iter().rev().collect();
-                        grouped.sort_by_key(|n| n.workspace_id);
-                        let mut last_ws_id: Option<u64> = None;
-                        for notif in &grouped {
-                            // Workspace section header
-                            if last_ws_id != Some(notif.workspace_id) {
-                                last_ws_id = Some(notif.workspace_id);
-                                if let Some(ws) =
-                                    self.workspaces.iter().find(|w| w.id == notif.workspace_id)
-                                {
-                                    ui.add_space(4.0);
-                                    ui.label(
-                                        egui::RichText::new(&ws.title)
-                                            .font(bold_font(11.0))
-                                            .color(egui::Color32::from_gray(120)),
-                                    );
-                                    ui.add_space(2.0);
-                                }
-                            }
-
-                            let response = ui.horizontal(|ui| {
-                                // Source icon + unread dot
-                                let source_icon = match notif.source {
-                                    NotificationSource::Bell => "\u{1f514}",  // 🔔
-                                    NotificationSource::Toast => "\u{1f4ac}", // 💬
-                                    NotificationSource::Cli => "\u{2328}",    // ⌨
-                                };
-                                let dot_color = if notif.read {
-                                    egui::Color32::from_gray(60)
-                                } else {
-                                    self.theme.chrome.accent
-                                };
-                                ui.label(
-                                    egui::RichText::new(source_icon).size(10.0).color(dot_color),
-                                );
-
-                                ui.vertical(|ui| {
-                                    let title = if notif.title.is_empty() {
-                                        &notif.body
-                                    } else {
-                                        &notif.title
-                                    };
-                                    ui.label(
-                                        egui::RichText::new(title)
-                                            .color(egui::Color32::from_gray(200)),
-                                    );
-                                    if !notif.title.is_empty() && !notif.body.is_empty() {
-                                        let body_display = if notif.body.len() > 100 {
-                                            format!("{}...", &notif.body[..97])
-                                        } else {
-                                            notif.body.clone()
-                                        };
-                                        ui.label(
-                                            egui::RichText::new(body_display)
-                                                .small()
-                                                .color(egui::Color32::from_gray(140)),
-                                        );
-                                    }
-                                    let age = render::format_duration(notif.created_at.elapsed());
-                                    ui.label(
-                                        egui::RichText::new(age)
-                                            .small()
-                                            .color(egui::Color32::from_gray(80)),
-                                    );
-                                });
-
-                                ui.with_layout(
-                                    egui::Layout::right_to_left(egui::Align::Center),
-                                    |ui| {
-                                        if ui.small_button("×").clicked() {
-                                            remove_id = Some(notif.id);
-                                        }
-                                    },
-                                );
-                            });
-                            if response.response.interact(egui::Sense::click()).clicked() {
-                                jump_to = Some((notif.workspace_id, notif.pane_id));
-                                close_panel = true;
-                            }
-                            ui.separator();
-                        }
-                    });
-                }
-            });
-
-        if mark_all {
-            self.notifications.mark_all_read();
-        }
-        if let Some(id) = remove_id {
-            self.notifications.remove_notification(id);
-        }
-        if let Some((ws_id, pane_id)) = jump_to {
-            if let Some(idx) = self.workspaces.iter().position(|ws| ws.id == ws_id) {
-                self.active_workspace_idx = idx;
-            }
-            self.set_focus(pane_id as PaneId);
-        }
-        if close_panel {
-            self.show_notification_panel = false;
-        }
-    }
-
     // --- Divider Drag ---
 
     fn handle_divider_drag(&mut self, ui: &egui::Ui, panel_rect: egui::Rect) {
@@ -3751,364 +1754,5 @@ impl AmuxApp {
         if primary_released {
             self.active_workspace_mut().dragging_divider = None;
         }
-    }
-
-    // --- Selection ---
-
-    fn copy_selection(&mut self) -> bool {
-        let focused = self.focused_pane_id();
-        let managed = match self.panes.get_mut(&focused) {
-            Some(m) => m,
-            None => return false,
-        };
-        let sel = match &managed.selection {
-            Some(s) => s.clone(),
-            None => return false,
-        };
-
-        let (cols, _) = managed.active_surface().pane.dimensions();
-        let (start, end) = sel.normalized();
-        let text =
-            selection::extract_selection_text(&managed.active_surface().pane, start, end, cols);
-
-        if text.is_empty() {
-            return false;
-        }
-
-        match arboard::Clipboard::new() {
-            Ok(mut clipboard) => {
-                if clipboard.set_text(&text).is_ok() {
-                    managed.selection = None; // Only clear on successful copy
-                }
-            }
-            Err(e) => {
-                tracing::warn!("Clipboard error: {}", e);
-            }
-        }
-        true
-    }
-
-    fn do_paste(&mut self, text: &str) {
-        let focused_id = self.focused_pane_id();
-        if let Some(managed) = self.panes.get_mut(&focused_id) {
-            let surface = managed.active_surface_mut();
-            surface.scroll_offset = 0;
-            surface.scroll_accum = 0.0;
-            if surface.pane.bracketed_paste_enabled() {
-                let _ = surface.pane.write_bytes(b"\x1b[200~");
-                let _ = surface.pane.write_bytes(text.as_bytes());
-                let _ = surface.pane.write_bytes(b"\x1b[201~");
-            } else {
-                let _ = surface.pane.write_bytes(text.as_bytes());
-            }
-        }
-    }
-
-    fn select_all_visible(&mut self) {
-        let focused = self.focused_pane_id();
-        let managed = match self.panes.get_mut(&focused) {
-            Some(m) => m,
-            None => return,
-        };
-        let surface = managed.active_surface();
-        let (cols, visible_rows) = surface.pane.dimensions();
-        let total = surface.pane.screen().scrollback_rows();
-        let scroll_offset = surface.scroll_offset;
-        let end_row = total.saturating_sub(scroll_offset);
-        let start_row = end_row.saturating_sub(visible_rows);
-
-        managed.selection = Some(SelectionState {
-            anchor: (0, start_row),
-            end: (cols.saturating_sub(1), end_row.saturating_sub(1)),
-            mode: SelectionMode::Cell,
-            active: false,
-        });
-    }
-
-    fn clear_selection_on_focused(&mut self) {
-        let focused = self.focused_pane_id();
-        if let Some(m) = self.panes.get_mut(&focused) {
-            m.selection = None;
-        }
-    }
-
-    // --- Selection Mouse ---
-
-    fn handle_selection_mouse(
-        &mut self,
-        ui: &egui::Ui,
-        pane_id: PaneId,
-        content_rect: egui::Rect,
-    ) -> bool {
-        let (cell_width, cell_height) = self.cell_dimensions(ui);
-
-        let managed = match self.panes.get(&pane_id) {
-            Some(m) => m,
-            None => return false,
-        };
-        let surface = managed.active_surface();
-        let (cols, visible_rows) = surface.pane.dimensions();
-        let total_rows = surface.pane.screen().scrollback_rows();
-        let scroll_offset = surface.scroll_offset;
-
-        let primary_pressed = ui.input(|i| i.pointer.primary_pressed());
-        let primary_down = ui.input(|i| i.pointer.primary_down());
-        let primary_released = ui.input(|i| i.pointer.primary_released());
-        let pointer_pos = ui.input(|i| i.pointer.hover_pos());
-
-        // Check if we're dragging a divider — skip selection if so
-        if self.active_workspace().dragging_divider.is_some() {
-            return false;
-        }
-
-        if primary_pressed {
-            if let Some(pos) = pointer_pos {
-                if !content_rect.contains(pos) {
-                    // Click outside content — clear any existing selection
-                    if let Some(m) = self.panes.get_mut(&pane_id) {
-                        m.selection = None;
-                    }
-                    return true;
-                }
-
-                let (col, stable_row) = selection::pointer_to_cell(
-                    pos,
-                    content_rect,
-                    cell_width,
-                    cell_height,
-                    scroll_offset,
-                    total_rows,
-                    visible_rows,
-                );
-                let col = col.min(cols.saturating_sub(1));
-
-                // Click count tracking
-                let now = Instant::now();
-                let dt = now.duration_since(self.last_click_time).as_millis();
-                let dist = (pos - self.last_click_pos).length();
-                if dt < 400 && dist < 5.0 {
-                    self.click_count = (self.click_count + 1).min(3);
-                } else {
-                    self.click_count = 1;
-                }
-                self.last_click_time = now;
-                self.last_click_pos = pos;
-
-                let (anchor, end, mode) = match self.click_count {
-                    2 => {
-                        // Word selection
-                        let text = selection::line_text_string(
-                            &managed.active_surface().pane,
-                            stable_row,
-                            cols,
-                        );
-                        let (wstart, wend) = selection::word_bounds_in_line(&text, col);
-                        (
-                            (wstart, stable_row),
-                            (wend, stable_row),
-                            SelectionMode::Word,
-                        )
-                    }
-                    3 => {
-                        // Line selection
-                        (
-                            (0, stable_row),
-                            (cols.saturating_sub(1), stable_row),
-                            SelectionMode::Line,
-                        )
-                    }
-                    _ => {
-                        // Cell selection
-                        ((col, stable_row), (col, stable_row), SelectionMode::Cell)
-                    }
-                };
-
-                if let Some(m) = self.panes.get_mut(&pane_id) {
-                    m.selection = Some(SelectionState {
-                        anchor,
-                        end,
-                        mode,
-                        active: true,
-                    });
-                }
-                return true;
-            }
-        } else if primary_down {
-            // Drag — update selection end
-            let has_active_selection = self
-                .panes
-                .get(&pane_id)
-                .and_then(|m| m.selection.as_ref())
-                .is_some_and(|s| s.active);
-
-            if has_active_selection {
-                if let Some(pos) = pointer_pos {
-                    let (col, stable_row) = selection::pointer_to_cell(
-                        pos,
-                        content_rect,
-                        cell_width,
-                        cell_height,
-                        scroll_offset,
-                        total_rows,
-                        visible_rows,
-                    );
-                    let col = col.min(cols.saturating_sub(1));
-
-                    if let Some(m) = self.panes.get_mut(&pane_id) {
-                        if let Some(ref mut sel) = m.selection {
-                            match sel.mode {
-                                SelectionMode::Cell => {
-                                    sel.end = (col, stable_row);
-                                }
-                                SelectionMode::Word => {
-                                    let text = selection::line_text_string(
-                                        &m.surfaces[m.active_surface_idx].pane,
-                                        stable_row,
-                                        cols,
-                                    );
-                                    let (_, wend) = selection::word_bounds_in_line(&text, col);
-                                    // Extend: keep anchor word start, update end word boundary
-                                    if stable_row > sel.anchor.1
-                                        || (stable_row == sel.anchor.1 && col >= sel.anchor.0)
-                                    {
-                                        sel.end = (wend, stable_row);
-                                    } else {
-                                        let (wstart, _) =
-                                            selection::word_bounds_in_line(&text, col);
-                                        sel.end = (wstart, stable_row);
-                                    }
-                                }
-                                SelectionMode::Line => {
-                                    if stable_row >= sel.anchor.1 {
-                                        sel.end = (cols.saturating_sub(1), stable_row);
-                                    } else {
-                                        sel.end = (0, stable_row);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    return true;
-                }
-            }
-        } else if primary_released {
-            if let Some(m) = self.panes.get_mut(&pane_id) {
-                if let Some(ref mut sel) = m.selection {
-                    sel.active = false;
-                    // If no actual drag (anchor == end), clear selection
-                    if sel.anchor == sel.end && sel.mode == SelectionMode::Cell {
-                        m.selection = None;
-                    }
-                }
-            }
-        }
-        false
-    }
-
-    // --- Input ---
-
-    fn handle_input(&mut self, ctx: &egui::Context) -> bool {
-        let events = ctx.input(|i| i.events.clone());
-        let focused_id = self.focused_pane_id();
-
-        // Clear selection when user types
-        let has_input = events.iter().any(|e| {
-            matches!(
-                e,
-                egui::Event::Text(_)
-                    | egui::Event::Paste(_)
-                    | egui::Event::Key { pressed: true, .. }
-            )
-        });
-        if has_input {
-            self.clear_selection_on_focused();
-        }
-
-        let managed = match self.panes.get_mut(&focused_id) {
-            Some(m) => m,
-            None => return has_input,
-        };
-        let surface = managed.active_surface_mut();
-
-        // When the process has exited, intercept Enter (close) and R (restart)
-        if surface.exited.is_some() {
-            let mut action = DeadPaneAction::None;
-            for event in &events {
-                if let egui::Event::Key {
-                    key,
-                    pressed: true,
-                    modifiers,
-                    ..
-                } = event
-                {
-                    match key {
-                        egui::Key::Enter => action = DeadPaneAction::Close,
-                        egui::Key::R if modifiers.is_none() => {
-                            action = DeadPaneAction::Restart;
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            match action {
-                DeadPaneAction::Close => self.close_pane(focused_id),
-                DeadPaneAction::Restart => {
-                    self.restart_surface(focused_id);
-                }
-                DeadPaneAction::None => {}
-            }
-            return has_input;
-        }
-
-        for event in &events {
-            match event {
-                egui::Event::Text(text) => {
-                    surface.scroll_offset = 0;
-                    surface.scroll_accum = 0.0;
-                    let _ = surface.pane.write_bytes(text.as_bytes());
-                }
-                egui::Event::Key {
-                    key,
-                    pressed: true,
-                    modifiers,
-                    ..
-                } => {
-                    if let Some(bytes) = key_encode::encode_egui_key(key, modifiers) {
-                        surface.scroll_offset = 0;
-                        surface.scroll_accum = 0.0;
-                        let _ = surface.pane.write_bytes(&bytes);
-                    }
-                }
-                egui::Event::Paste(text) => {
-                    surface.scroll_offset = 0;
-                    surface.scroll_accum = 0.0;
-                    if surface.pane.bracketed_paste_enabled() {
-                        let _ = surface.pane.write_bytes(b"\x1b[200~");
-                        let _ = surface.pane.write_bytes(text.as_bytes());
-                        let _ = surface.pane.write_bytes(b"\x1b[201~");
-                    } else {
-                        let _ = surface.pane.write_bytes(text.as_bytes());
-                    }
-                }
-                egui::Event::Ime(ime_event) => match ime_event {
-                    egui::ImeEvent::Commit(text) => {
-                        surface.scroll_offset = 0;
-                        surface.scroll_accum = 0.0;
-                        self.ime_preedit = None;
-                        let _ = surface.pane.write_bytes(text.as_bytes());
-                    }
-                    egui::ImeEvent::Preedit(text) => {
-                        self.ime_preedit = if text.is_empty() {
-                            None
-                        } else {
-                            Some(text.clone())
-                        };
-                    }
-                    _ => {}
-                },
-                _ => {}
-            }
-        }
-        has_input
     }
 }
